QEMU with some utilities
The normal qemu-git package, maintained by @ajs124, is missing kvm_stat,
which is the particular utility I need right this minute.

Also, this PKGBUILD seems to NOT do a split package of the libacard, as
well it uses the SeaBIOS shipped with QEMU itself. I don't know enough
about QEMU + kvm yet to know whether this is reasonable or not.